+++ package/src/lib.rs
@@ -49,7 +49,7 @@ fn run_batch(inputs: &[String], options: &Option<breadchunks::ChunkOptions>) -> 
                     headers: c.headers,
                     breadcrumb: c.breadcrumb,
                     text: c.text,
-                    length: c.length,
+                    length: c.length as u32, // usize→u32 narrowing for napi; docs >4 GiB unsupported on Node binding
                 })
                 .collect()
         })
